+++ src/main/java/com/revature/Flumblr/repositories/CommentRepository.java
@@ -3,7 +3,7 @@ package com.revature.Flumblr.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+// import java.util.List;
 
 /**
  * The PostRepository interface provides database operations for Post entities.
