+++ src/main/java/com/revature/Flumblr/services/PostService.java
@@ -1,6 +1,8 @@
 package com.revature.Flumblr.services;
 
+import com.revature.Flumblr.repositories.CommentRepository;
 import com.revature.Flumblr.repositories.PostRepository;
+import com.revature.Flumblr.repositories.UserRepository;
 import com.revature.Flumblr.utils.custom_exceptions.PostNotFoundException;
 import java.util.List;
 import java.util.Optional;
@@ -10,11 +12,14 @@ import java.util.ArrayList;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+
+import com.revature.Flumblr.dtos.requests.NewCommentRequest;
 import com.revature.Flumblr.dtos.responses.PostResponse;
 import com.revature.Flumblr.entities.User;
 
 import lombok.AllArgsConstructor;
 
+import com.revature.Flumblr.entities.Comment;
 import com.revature.Flumblr.entities.Follow;
 import com.revature.Flumblr.entities.Post;
 
@@ -23,6 +28,9 @@ import com.revature.Flumblr.entities.Post;
 public class PostService {
     private final PostRepository postRepo;
     private final UserService userService;
+       private final CommentRepository commentRepository;
+   private final UserRepository userRepository;
+   private final PostRepository postRepository;
 
     public List<PostResponse> getFeed(String userId, int page) {
         User user = userService.getUserById(userId);
@@ -53,4 +61,11 @@ public class PostService {
         if(userPost.isEmpty()) throw new PostNotFoundException(postId);
         return new PostResponse(userPost.get());
     }
+
+        public void commentOnPost(NewCommentRequest req) {
+        User user = userRepository.getReferenceById(req.getUser_id());
+        Post post = postRepository.getReferenceById(req.getPost_id());
+        Comment com = new Comment(req.getComment(), post, user);
+        commentRepository.save(com);     
+    }   
 }
